Riferimenti

Questa pagina ha riferimenti a tutti i tipi e funzioni utilizzati.

Tipi

TGW3D.ChainOp — Type

ChainOp = SparseArrays.SparseMatrixCSC{Int8,Int}

Alias declation of LAR-specific data structure. SparseMatrix in Compressed Sparse Column format, contains the coordinate representation of an operator between linear spaces of P-chains. Operators $P-Boundary : P-Chain -> (P-1)-Chain$ and $P-Coboundary : P-Chain -> (P+1)-Chain$ are typically stored as ChainOp with elements in ${-1,0,1}$ or in ${0,1}$, for signed and unsigned operators, respectively.

TGW3D.Points — Type

Points = Array{Number,2}

Alias declation of LAR-specific data structure. Dense Array{Number,2,1} $M x N$ to store the position of vertices (0-cells) of a cellular complex. The number of rows $M$ is the dimension of the embedding space. The number of columns $N$ is the number of vertices.

Funzioni

TGW3D.frag_face — Method

fragface( V::Points, EV::ChainOp, FE::ChainOp, spidx::Vector{Int64}, sigma::Int64)

Prende la faccia sigma e la trasforma in 2D per poter calcolare le intersezioni con le facce in sp_idx[sigma] ed ottenere la disposizione 2D della faccia sigma.

TGW3D.frag_face_channel — Method

fragfacechannel( inchan, outchan, V::Points, EV::ChainOp, FE::ChainOp, sp_idx)

Funziona che parallelizza, con l'utilizzo dei canali, la frammentazione delle facce in FE rispetto le facce in sp_idx.

TGW3D.merge_vertices — Function

merge_vertices( V::Points, EV::ChainOp, FE::ChainOp, [err=1e-4])

Rimuove i vertici congruenti ad un singolo rappresentatante, traduce i lati per tener conto della congruenza ed otteniene nuove facce congruenti.

TGW3D.removeinnerloops — Method

removeinnerloops(
    g::Int64, 
    nFE::ChainOp)

Rimuove le facce all'interno dei cicli interni dalla matrice sparsa nFE. Il valore restituito ha g righe in meno rispetto all'input nFE.

TGW3D.spatial_arrangement — Function

spatial_arrangement(V::Points, copEV::ChainOp, copFE::ChainOp; [multiproc::Bool])

Compute the arrangement on the given cellular complex 2-skeleton in 3D.

A cellular complex is arranged when the intersection of every possible pair of cell of the complex is empty and the union of all the cells is the whole Euclidean space. The function returns the full arranged complex as a list of vertices V and a chain of borders EV, FE, CF.

Additional arguments:

multiproc::Bool: Runs the computation in parallel mode. Defaults to false.

TGW3D.spatial_arrangement_1 — Function

function spatialarrangement1( V::Points, copEV::ChainOp, copFE::ChainOp, [multiproc::Bool=false])

Si occupa del processo di frammentazione delle facce per l’utilizzo del planar arrangement. Richiama le funzioni frag_face e `merge_vertices' per ritornare i nuovi vertici, lati e facce.

TGW3D.spatial_arrangement_2 — Function

function spatialarrangement2( rV::Points, rcopEV::ChainOp, rcopFE::ChainOp, [multiproc::Bool=false])

TODO