TGW3D.TGW3D — Module

L’algoritmo Topological Gift Wrapping calcola le d-celle di una partizione di spazio generate da loro partendo da un oggetto geometrico d-1 dimensionale.

TGW prende una matrice sparsa di dimensione d-1 in input e produce in output la matrice sparsa di dimensione d sconosciuta aumentata dalle celle esterne.

Riferimenti

Questa pagina ha i riferimenti a tutti i tipi, metodi e funzioni utilizzati.

Tipi

TGW3D.ChainOp — Type

ChainOp = SparseArrays.SparseMatrixCSC{Int8,Int}

Dichiarazione Alias di specifiche strutture dati di LAR. SparseMatrix nel formato Colonne sparse compresse, contiene la rappresentazione in coordinate di un operatore tra lo spazio lineare delle P-chains. Operatori $P-Boundary : P-Chain -> (P-1)-Chain$ e $P-Coboundary : P-Chain -> (P+1)-Chain$ sono tipicamente immagazinati come ChainOpcon elementi in ${-1,0,1}$ oppure in ${0,1}$, per operatori assegnati e non-assegnati rispettivamente.

TGW3D.Points — Type

Points = Array{Number,2}

Dichiarazione Alias di specifiche strutture dati di LAR. Array{Number,2,1} $M x N$ compatto per immagazzinare la posizione dei vertici (0-cells) di un complesso cellulare. Il numero delle righe $M$ è la dimensione dello spazio di inclusione. Il numero delle colonne $N$ è il numero dei vertici.

Funzioni

TGW3D.frag_face — Method

function frag_face(
        V::Points, 
        EV::ChainOp, 
        FE::ChainOp, 
        sp_idx::Vector{Int64}, 
        sigma::Int64)

Prende la faccia sigma e la trasforma in 2D per poter calcolare le intersezioni con le facce in sp_idx[sigma] ed ottenere la disposizione 2D della faccia sigma.

TGW3D.frag_face_channel — Method

function frag_face_channel(
        in_chan, 
        out_chan, 
        V::Points, 
        EV::ChainOp, 
        FE::ChainOp, 
        sp_idx::Vector{Int64})

Funzione che parallelizza, con l'utilizzo dei canali, la frammentazione delle facce in FE rispetto le facce in sp_idx.

TGW3D.merge_vertices — Function

function merge_vertices(
        V::Points, 
        EV::ChainOp, 
        FE::ChainOp, 
        [err=1e-4])

Rimuove i vertici congruenti ad un singolo rappresentatante, traduce i lati per tener conto della congruenza ed otteniene nuove facce congruenti.

Argomenti addizionali:

err: Limite di errore massimo che si vuole utilizzare. Di Defaults a 1e-4.

TGW3D.removeinnerloops — Method

function removeinnerloops(
        g::Int64, 
        nFE::ChainOp)

Rimuove le facce all'interno dei cicli interni dalla matrice sparsa nFE. Il valore restituito ha g righe in meno rispetto all'input nFE.

TGW3D.spatial_arrangement — Function

function spatial_arrangement(
        V::Points, 
        copEV::ChainOp, 
        copFE::ChainOp; 
        [multiproc::Bool])

Calcola la disposizione sulle cellule complesse 2-skeleton date in 3D.

Un complesso cellulare è disposto quando l'intersezione di ogni coppia di celle del complesso è vuota e l'unione di tutte le celle rappresenta l'intero spazio Euclideo. La funzione ritorna la piena disposizione complessa come una lista di vertici V e una catena di lati EV, FE, CF.

Argomenti addizionali:

multiproc::Bool: Esegue la computazione in modalità parallela. Di Defaults a false.

TGW3D.spatial_arrangement_1 — Function

function spatial_arrangement_1(
        V::Points,
        copEV::ChainOp,
        copFE::ChainOp, 
        [multiproc::Bool=false])

Si occupa del processo di frammentazione delle facce per l'utilizzo del planar arrangement. Richiama le funzioni frag_face e `merge_vertices' per ritornare i nuovi vertici, lati e facce.

Argomenti addizionali:

multiproc::Bool: Esegue la computazione in modalità parallela. Di Defaults a false.

TGW3D.spatial_arrangement_2 — Function

function spatial_arrangement_2(
        rV::Points, 
        rcopEV::ChainOp, 
        rcopFE::ChainOp, 
        [multiproc::Bool=false])

Effettua la ricostruzione delle facce permettendo il wrapping spaziale 3D.

Argomenti addizionali:

multiproc::Bool: Esegue la computazione in modalità parallela. Di Defaults a false.